+++ src/main/java/ladder/server/domain/ladder/ConnectionJudgement.java
@@ -1,4 +1,4 @@
-package ladder.server.domain;
+package ladder.server.domain.ladder;
 
 /**
  * 사다리가 연결되는지, 연결되지 않는지를 판단하는 인터페이스
