+++ src/main/java/ladder/server/domain/ladder/RandomConnectionJudgement.java
@@ -1,4 +1,4 @@
-package ladder.server.domain;
+package ladder.server.domain.ladder;
 
 import java.util.concurrent.ThreadLocalRandom;
