+++ src/main/java/ladder/server/domain/player/PlayerName.java
@@ -1,4 +1,4 @@
-package ladder.server.domain;
+package ladder.server.domain.player;
 
 /**
  * 참가자의 이름을 검증하는 역할을 가진 클래스
