+++ src/main/java/ladder/server/domain/result/ItemName.java
@@ -1,4 +1,4 @@
-package ladder.server.domain;
+package ladder.server.domain.result;
 
 /**
  * 결과의 이름을 검증하는 역할을 가진 클래스
