+++ src/test/java/ladder/server/domain/ladder/MockConnectionJudgement.java
@@ -1,6 +1,4 @@
-package ladder.server.domain;
-
-import ladder.server.domain.ConnectionJudgement;
+package ladder.server.domain.ladder;
 
 /**
  * 원하는대로 연결 혹은 연결되지 않음을 판단하는 클래스
