+++ src/test/java/ladder/server/domain/ladder/PointTest.java
@@ -1,8 +1,9 @@
-package ladder.server.domain;
+package ladder.server.domain.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import ladder.server.domain.common.Position;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
