+++ src/test/java/ladder/server/domain/player/PlayerTest.java
@@ -1,9 +1,8 @@
-package ladder.server.domain;
+package ladder.server.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import ladder.server.domain.Player;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
