+++ src/test/java/ladder/server/domain/result/ItemTest.java
@@ -1,10 +1,9 @@
-package ladder.server.domain;
+package ladder.server.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import ladder.server.domain.Item;
-import ladder.server.domain.Position;
+import ladder.server.domain.common.Position;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
