+++ src/test/java/ladder/server/domain/result/ResultTest.java
@@ -1,12 +1,11 @@
-package ladder.server.domain;
+package ladder.server.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import ladder.server.domain.Position;
-import ladder.server.domain.Result;
+import ladder.server.domain.common.Position;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
